Refactor, split general Graph and ModelGraph stuff.
